shiny/requirements.txt
shiny>=0.4.0
plotly>=5.13.0
pandas>=1.3.5
shinywidgets
boto3
wordcloud
python-dotenv>=0.15.0,<1.0.0
